make schedule section a section
I should not try to write HTML when I'm tired, it turns out. Because schedule was templated as a page-header, the CSS demanded that it have a more prominent z-axis than the rest of the content and was covering it up. Thanks to @fmichonneau for pointing out the cause: https://github.com/zkamvar/sandpaper-docs/issues/1