attempt to fix heading sizes
The default headings made no difference between mobile and tablet views
and were set for the former. The H1 and H2 elements were reset to be
upgraded on the desktop view, but H3 headings were not changes and H4
headings were not at all touched, which lead to awkward situations in
which headings that were of lower level appeared to be a higher level.